% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqmagic-package.R
\docType{package}
\name{dqmagic-package}
\alias{dqmagic}
\alias{dqmagic-package}
\title{R interface to libmagic}
\description{
The dqmagic package provides an R interface for the
\href{https://linux.die.net/man/3/libmagic}{\code{libmagic}} file type
identification library similar to the Unix
\href{https://linux.die.net/man/1/file}{\code{file}} command. This allows for
file type identification based on a file's content instead of it's extension.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.daqana.org/dqmagic/}
  \item Report bugs at \url{https://github.com/daqana/dqmagic/issues}
}

}
\author{
\strong{Maintainer}: Ralf Stubner \email{ralf.stubner@daqana.com}

}
