% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{file_type}
\alias{file_type}
\title{Get file types}
\usage{
file_type(files)
}
\arguments{
\item{files}{Vector of file names including full or relative paths}
}
\value{
Vector of textual descriptions of file type
}
\description{
Get file types.
}
