% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{magicPathDefault}
\alias{magicPathDefault}
\title{Get default magic}
\usage{
magicPathDefault()
}
\value{
Paths of the default magic files and directories
}
\description{
Get paths for the default magic files and directories.
}
