% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{magicVersion}
\alias{magicVersion}
\title{Get libmagic version}
\usage{
magicVersion()
}
\value{
Integer representing the version number
}
\description{
Get version of the libmagic library as integer: MAJOR * 100 + MINOR.
}
