% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqmagic-package.R
\docType{package}
\name{dqmagic-package}
\alias{dqmagic}
\alias{dqmagic-package}
\title{R interface to libmagic}
\description{
The dqmagic package provides an R interface for libmagic.
}
\author{
\strong{Maintainer}: Ralf Stubner \email{ralf.stubner@r-institute.com}

}
