% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{file_type}
\alias{file_type}
\title{Get file type}
\usage{
file_type(file)
}
\arguments{
\item{file}{Name of the file including full or relative path}
}
\value{
Textual description of file type
}
\description{
Get file type.
}
