% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{file_type}
\alias{file_type}
\title{Get file types}
\usage{
file_type(files, mime_type = FALSE, mime_encoding = FALSE)
}
\arguments{
\item{files}{Vector of file names including full or relative paths}

\item{mime_type}{Logical}

\item{mime_encoding}{Logical}
}
\value{
Vector of textual descriptions or MIME types/encodings of file type
}
\description{
Get file types.
}
