% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{file_type}
\alias{file_type}
\title{Get file types}
\usage{
file_type(files, mime_type = FALSE, mime_encoding = FALSE,
  magicfiles = NULL)
}
\arguments{
\item{files}{Vector of file names including full or relative paths}

\item{mime_type}{Logical indicating whether the MIME type should be returned}

\item{mime_encoding}{Logical indicating whether the MIME encoding should be returned}

\item{magicfiles}{Vector with file or directory names to be used as magic file; the first
usable location is selected}
}
\value{
Vector of textual file type descriptions or MIME types/encodings
}
\description{
Get file types based on the file's content.
}
